Move PaymentMethodParams `label` to  where it belongs and re-implement it to reuse the localized type displayName
