+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodParams.swift
@@ -1343,86 +1343,3 @@ extension STPPaymentMethodParams {
         }
     }
 }
-
-extension STPPaymentMethodParams {
-    @objc public var label: String {
-        switch type {
-        case .alipay:
-            return "Alipay"  // ? Why aren't these localized?
-        case .card:
-            if let card = card {
-                let brand = STPCardValidator.brand(forNumber: card.number ?? "")
-                let brandString = STPCardBrandUtilities.stringFrom(brand)
-                return "\(brandString ?? "") \(card.last4 ?? "")"
-            } else {
-                return STPCardBrandUtilities.stringFrom(.unknown) ?? ""
-            }
-        case .iDEAL:
-            return "iDEAL"
-        case .FPX:
-            if let fpx = fpx {
-                return STPFPXBank.stringFrom(fpx.bank) ?? ""
-            } else {
-                return "FPX"
-            }
-        case .SEPADebit:
-            return "SEPA Debit"
-        case .bacsDebit:
-            return "Bacs Debit"
-        case .AUBECSDebit:
-            return "AU BECS Debit"
-        case .giropay:
-            return "giropay"
-        case .przelewy24:
-            return "Przelewy24"
-        case .EPS:
-            return "EPS"
-        case .bancontact:
-            return "Bancontact"
-        case .netBanking:
-            return "NetBanking"
-        case .OXXO:
-            return "OXXO"
-        case .sofort:
-            return "Sofort"
-        case .UPI:
-            return "UPI"
-        case .grabPay:
-            return "GrabPay"
-        case .payPal:
-            return "PayPal"
-        case .afterpayClearpay:
-            return "Afterpay Clearpay"
-        case .blik:
-            return "BLIK"
-        case .weChatPay:
-            return "WeChat Pay"
-        case .boleto:
-            return "Boleto"
-        case .link:
-            return "Link"
-        case .klarna:
-            return "Klarna"
-        case .affirm:
-            return "Affirm"
-        case .USBankAccount:
-            return "US Bank Account"
-        case .cashApp:
-            return "Cash App Pay"
-        case .revolutPay:
-            return "Revolut Pay"
-        case.twint:
-            return "TWINT"
-        case .multibanco:
-            return "Multibanco"
-        case .cardPresent, .unknown:
-            return STPLocalizedString("Unknown", "Default missing source type label")
-        case .paynow, .zip, .amazonPay, .alma, .mobilePay, .konbini, .promptPay, .swish,
-             .sunbit, .billie, .satispay:
-            // Use the label already defined in STPPaymentMethodType; the params object for these types don't contain additional information that affect the display label (like cards do)
-            return type.displayName
-        @unknown default:
-            return STPLocalizedString("Unknown", "Default missing source type label")
-        }
-    }
-}
